fixed misreported device id

+++ src/database.h
@@ -26,8 +26,8 @@ void DB_connect();
 void DB_query(char * item ...);
 char* DB_resultAsText();
 bool DB_isQueryReady();
-void DB_addGpsPacket(int deviceId, int status, char* lat, char* latRef,
+void DB_addGpsPacket(uint16_t deviceId, int status, char* lat, char* latRef,
     char* lon, char* lonRef, char* spd, char* hdg);
-void DB_addDataPacket(int deviceId, int DI, int altitude, int rate, uint16_t analog[18]);
+void DB_addDataPacket(uint16_t deviceId, int DI, int altitude, int rate, uint16_t analog[18]);
 
 #endif
